Tree: Move visitDelta
