+++ packages/dds/tree/src/changeset/index.ts
@@ -9,4 +9,3 @@
 
 export * from "./format";
 export * from "./toDelta";
-export * from "./visit";
