+++ packages/dds/tree/src/tree/index.ts
@@ -18,6 +18,7 @@ export {
 export * from "./pathTree";
 export * from "./anchorSet";
 export * from "./treeTextFormat";
+export * from "./visitDelta";
 
 // Split this up into separate import and export for compatibility with API-Extractor.
 import * as Delta from "./delta";
