+++ packages/dds/tree/src/tree/visitDelta.ts
@@ -3,8 +3,9 @@
  * Licensed under the MIT License.
  */
 
-import { FieldKey, Value, Delta } from "../tree";
 import { fail, unreachableCase } from "../util";
+import { FieldKey, Value } from "./types";
+import * as Delta from "./delta";
 
 /**
  * Implementation notes:
